Changing components structure
