+++ src/components/Header.jsx
@@ -1,13 +1,10 @@
-import { PodcastListProvider } from "../context/filterContext";
 import InputFilter from "./InputFilter";
 
 const Header = ({ setFiltered, filtered }) => {
   return (
     <div className="header">
       <h2 className="title">Podcaster</h2>
-      <PodcastListProvider>
-        <InputFilter setFiltered={setFiltered} filtered={filtered} />
-      </PodcastListProvider>
+      <InputFilter setFiltered={setFiltered} filtered={filtered} />
     </div>
   );
 };
