+++ src/components/InputFilter.jsx
@@ -1,40 +1,7 @@
-import { useState, useEffect } from "react";
-import { usePodcastListContext } from "../context/filterContext";
-import podcastList from "../services/podcastList";
+import ListCounter from "./ListCounter";
+import PodcastCardsFiltered from "./PodcastCardsFiltered";
 
-const InputFilter = ({ setFiltered, filtered }) => {
-  debugger;
-  const [error, setError] = useState(null);
-  const [filterText, setFilterText] = useState("");
-  const { podcastListFiltered, setPodcastListFiltered } = usePodcastListContext();
-
-  useEffect(() => {
-    const podcastFilteredArray = podcastList.filter((element) => {
-      let elementName = element.name;
-      let authorName = element.artist;
-
-      if (filterText) {
-        if (
-          elementName.toLowerCase().includes(filterText.toLowerCase()) ||
-          authorName.toLowerCase().includes(filterText.toLowerCase())
-        ) {
-          return element;
-        }
-      }
-    });
-    setPodcastListFiltered(podcastFilteredArray);
-  }, [filterText]);
-
-  useEffect(() => {
-    if (podcastListFiltered) {
-      if (filterText && podcastListFiltered.length === 0) {
-        setError(true);
-      } else {
-        setError(false);
-      }
-    }
-  }, [filterText]);
-
+const InputFilter = ({ setFiltered, filtered, setFilterText }) => {
   const handleFilterChange = (event) => {
     setFilterText(event.target.value);
     setFiltered(true);
@@ -46,14 +13,7 @@ const InputFilter = ({ setFiltered, filtered }) => {
   return (
     <>
       <input type="text" placeholder="Search podcast" onChange={handleFilterChange} />
-      {(error === true && <div>No matching data</div>) || (
-        <h4>
-          {podcastListFiltered &&
-            (podcastListFiltered.length === 0 || error === null
-              ? podcastList.length
-              : podcastListFiltered.length)}
-        </h4>
-      )}
+      <ListCounter filtered={filtered} />
     </>
   );
 };
