+++ src/components/ListCounter.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ListCounter = ({ error }) => {
+  return <div>ListCounter</div>;
+};
+
+export default ListCounter;
